third_party/blink/web_tests/http/tests/origin_trials/webexposed/declarative-shadow-dom-origin-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py --version 3 http://127.0.0.1:8000 DeclarativeShadowDOM --expire-timestamp=2000000000
-- -->
<meta http-equiv="origin-trial" content="A80LutZa5Y3GLQ8KvEiYH5ism5H62hH0Yk231NOhuQlkVDIMoJ8lh9xPALEPNZuWT+JymA4H7CkeOeystF+K9gwAAABceyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiRGVjbGFyYXRpdmVTaGFkb3dET00iLCAiZXhwaXJ5IjogMjAwMDAwMDAwMH0=" />
<title>DeclarativeShadowDOM - enabled by origin trial</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(t => {
  // Basic feature detection
  assert_true(HTMLTemplateElement.prototype.hasOwnProperty("shadowRoot"),'Unable to feature detect');
  // Basic feature use
  const div = document.createElement('div');
  document.allowDeclarativeShadowDom = true;
  div.innerHTML = '<div id="host"><template shadowroot="open"></template></div>';
  const host = div.querySelector('#host');
  assert_true(!!host.shadowRoot,"No shadow root found");
  document.allowDeclarativeShadowDom = false; // For other tests
}, 'Declarative Shadow DOM enabled for origin-trial enabled document.');
</script>

<script>
async_test((t) => {
  window.addEventListener('message', (e) => {
    t.step(() => {
      assert_true(e.data.gotHost);
      assert_true(e.data.gotShadow);
      t.done();
    });
  });
}, 'Origin trial enables the allow-declarative-shadow-dom sandbox flag');
</script>

<iframe id="iframe" sandbox="allow-same-origin allow-scripts allow-declarative-shadow-dom" src="resources/declarative-child-frame.html"></iframe>
